LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-007-expected.txt


PASS Should use the last position option initially
PASS Should stay at initial position, since it still fits
FAIL Should use the third position option with enough space left assert_equals: Anchored element should be at the left of anchor expected 586 but got 786
FAIL Should use the first position option with enough space left and below assert_equals: Anchored element should be at the left of anchor expected 135 but got 335
FAIL Should use the second position option with enough space below (and not enough above) assert_equals: Anchored element should be at the bottom of anchor expected 199 but got -1
